refactor: adds useMutateInstallation hook to manage state hooks in CreateInstallation and UpdateInstallation components
